flakes: use stable release branches for `nixpkgs` inputs
This is the best practice for Flakes or clients of nixpkgs in general.

The `release-<YY.MM>` branch is actually a dev branch:
Hydra continually tests the tip of this branch.
When the tests succeed, the `nixos-<YY.MM>` release branch is updated to
point to the succeeding commit.

`release-<YY.MM>` is to `nixos-<YY.MM>` what `master` is to `nixpkgs-unstable`.

See also: https://github.com/NixOS/nixpkgs/blob/e88fb997e32982a5f37b933583e618265a32d85d/CONTRIBUTING.md#flow-of-merged-pull-requests